bcolors moved into __init__.py
